+++ CHANGES.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+- Add `getText` and `getJSON` response middleware for easily parsing and
+  returning response content in one shot
 - Build everything into the package root
 
 ## [v4.1.0]